Extract `Prism::Translator` class
