+++ parser/prism/BUILD
@@ -2,9 +2,11 @@ cc_library(
     name = "prism",
     srcs = [
         "Parser.cc",
+        "Translator.cc",
     ],
     hdrs = [
         "Parser.h",
+        "Translator.h",
     ],
     linkstatic = select({
         "//tools/config:linkshared": 0,
@@ -13,5 +15,6 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//core",
+        "//parser", # Legacy parser, needed for translating to its nodes defined in `parser/Node_gen.h`
     ],
 )
